Move flask form out to its own file
