+++ quaff/strategies/__init__.py
@@ -1,2 +1,3 @@
 from .flask_endpoint import FlaskEndpoint
+from .flask_form import FlaskForm
 from .command_args import CommandArgs
